# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Asset Management System",
        "contact": {
            "email": "user@email.com"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Asset Management System API"
        }
    ],
    "paths": {
        "/assets": {
            "get": {
                "tags": [
                    "Assets"
                ],
                "summary": "Get a list of assets",
                "operationId": "839a40ad0082bc4601e4e5b4d7e8d3d7",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Assets"
                ],
                "summary": "Create new asset",
                "operationId": "ef1f9ce040e398e44a16cc8feaba324f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAssetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/assets/{id}": {
            "delete": {
                "tags": [
                    "Assets"
                ],
                "summary": "Delete a assets",
                "operationId": "dc46a52e31fd7e6c2892a6fc7ae4d39f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the asset to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Asset deleted successfully"
                    },
                    "404": {
                        "description": "Asset not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Assets"
                ],
                "summary": "Update asset",
                "operationId": "71af9c890fbdbd3d02f25c3c90f3421f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the asset to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAssetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreateAssetRequest": {
                "required": [
                    "name",
                    "status"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Iphone SE"
                    },
                    "status": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "UpdateAssetRequest": {
                "required": [
                    "name",
                    "status"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Iphone SE"
                    },
                    "status": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Assets",
            "description": "Assets"
        }
    ]
}